Add loading animation for signing URL and Google doc viewer for viewing documents
When viewing a file, added a loading animation to display in the new tab while the server is signing and returning the signed URL. Added Google docs viewer to view files depending on their type. If the file type is 'document' it will be opened in google docs viewer. This involved setting the file type when uploading (depending on the file extension), storing this in the DB and returning the type when signing the URL for viewing. This means only newly uploaded files will display correctly since all file types were originally 'document'. Also added a generic file type called 'file' in the File model.