JWT payload becoming invalid fix
Added date modified and date created field to the user model, added an api request to get the last modified date, added code to the isLoggedIn function in the authentication service to check the jwt and last modified date whenever called (whenever changing views/pages) - compares the the jwt iat date and the last modified date, if out of date token a message will be displayed and the user will be logged out. Also added a message for the expiry date.