Map page search function
Linked the file search results to the markers on the map, if a user hovers the a search results a circle is displayed around the marker, if the user clicks a search results the map is panned and zoomed to the marker and the markers popup box is opened. Added more styling. Minor bug with the css for the cross in the search input.