File model & controller additions, marker link model & controller edits
Changed the marker links model to use references for the creator and the precedent and dependent files, editted the controller to populate the response with what is needed, some front end modifications to pass and accept this data correctly. Added a markerLinks array field to the File model that references marker links and added a function to the deleteFile function to delete any marker links associated with the file.