File browser modifications
Better sorting, fixed delete on the front-end. deleteFileS3 in the backend wasn't deleting files on S3 successfully (S3 returns success even if the key it's deleting does not exist). The file browser page was sending it's key as an object, not a string. The route in index also needed (*). Appears to be working now but will probably need fixing across the front-end. #23