Added forgot password functionality
Entering an email into forgot password checks that email then generates a random 20 character long string which is saved into the database, along with the datetime + 1 hour. An email is sent to the users email address with a link with the unique string as a url param. This link opens a page where the user enters their new password (HTML + CSS needed), the database then checks the token and time and if okay adds the new password. Basics implemented but some error checking needed and a lot of front-end work.