Survey Work - files renamed + more
Renamed some files to keep consistent, added a couple api calls for checking if the survey exists and for reading individual survey responses. The check survey function checks if the survey exists when a user enters the access id on the complete survey page and checks the users IP address against the existing survey responses for that survey, if IP address exists and the response date is within 5 minutes the user is denied access.