Added custom map marker icons to the map page
Also small edit to the file model, removed 'document' type and replaced with 'doc' and 'pdf'. This means files in local databases will need to be re-added or the type field in the your database will need to be editted for each file to display the correct map markers. Some front end modifications still need to be made becuase of this such as the view file function on a few pages, this is becuase the file is opened in google docs if its type is 'document', this needs to be editted to 'doc' or 'pdf' now, and perhaps 'text' too.