NPM UPDATE: Test libraries
Been working on setting up the unit testing, started testing the test libraries on the analysis service, the analysis controller, and the knowledgebase controller. Starting to get the hang of it but some code needs to be modified for it to work 100% correctly. Made some system wide modifications, tested most things and appears to be okay - if you notice any new errors let me know.