Back-end modifications to update analysisresults and files
Added reference to files for the analyses and references to the analyses for the files. If a file belongs to an anlysis and is deleted (on the browser page) a second confirmation message will appear. If deleted, the information for that file is added to the anlysis document in the database and the file is marked as deleted in the file list on the visulisation page (instead of just disappearing).