Add date range filter for the map
Filters map markers and map links based on the date set by the date slider. Each step or 'update' all the map markers are removed and the markers that are within the date range are re-added. May be too computationally intense, so maybe we should change the event listener to update on 'change', which occurs when the slider is dropped, not dragged.